+++ tools/check/check_code_quality.sh
@@ -75,11 +75,14 @@ echo
 echo "Search for forbidden patterns in resources..."
 
 ${searchForbiddenStringsScript} ./tools/check/forbidden_strings_in_resources.txt \
-    ./vector/src/main/res/layout \
+    ./vector/src/main/res/anim \
+    ./vector/src/main/res/color \
     ./vector/src/main/res/menu \
+    ./vector/src/main/res/layout \
     ./vector/src/main/res/values \
     ./vector/src/main/res/values-v21 \
-    ./vector/src/main/res/values-w820dp
+    ./vector/src/main/res/values-w820dp \
+    ./vector/src/main/res/xml
 
 resultForbiddenStringInResource=$?