+++ vector/src/main/java/im/vector/activity/AccountCreationCaptchaActivity.java
@@ -41,6 +41,7 @@ import org.matrix.androidsdk.util.Log;
 import java.net.URLDecoder;
 import java.util.Formatter;
 import java.util.HashMap;
+import java.util.Map;
 
 import im.vector.R;
 
@@ -197,7 +198,7 @@ public class AccountCreationCaptchaActivity extends RiotAppCompatActivity {
             public boolean shouldOverrideUrlLoading(android.webkit.WebView view, java.lang.String url) {
                 if ((null != url) && url.startsWith("js:")) {
                     String json = url.substring(3);
-                    HashMap<String, String> parameters = null;
+                    Map<String, String> parameters = null;
 
                     try {
                         // URL decode