+++ vector/src/main/java/im/vector/activity/FallbackLoginActivity.java
@@ -37,6 +37,7 @@ import org.matrix.androidsdk.util.Log;
 
 import java.net.URLDecoder;
 import java.util.HashMap;
+import java.util.Map;
 
 import im.vector.R;
 
@@ -181,7 +182,7 @@ public class FallbackLoginActivity extends RiotAppCompatActivity {
             public boolean shouldOverrideUrlLoading(android.webkit.WebView view, java.lang.String url) {
                 if ((null != url) && url.startsWith("js:")) {
                     String json = url.substring(3);
-                    HashMap<String, Object> serverParams = null;
+                    Map<String, Object> serverParams = null;
 
                     try {
                         // URL decode