+++ vector/src/main/java/im/vector/activity/VectorSharedFilesActivity.java
@@ -18,16 +18,15 @@
 package im.vector.activity;
 
 import android.content.Intent;
-import android.os.Bundle;
-
-import org.matrix.androidsdk.data.RoomMediaMessage;
-import org.matrix.androidsdk.util.Log;
 
 import org.matrix.androidsdk.MXSession;
+import org.matrix.androidsdk.data.RoomMediaMessage;
 import org.matrix.androidsdk.util.ContentUtils;
+import org.matrix.androidsdk.util.Log;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.List;
 
 import im.vector.Matrix;
 import im.vector.R;
@@ -119,7 +118,7 @@ public class VectorSharedFilesActivity extends RiotAppCompatActivity {
 
         sharedFolder.mkdir();
 
-        ArrayList<RoomMediaMessage> cachedFiles = new ArrayList<>(RoomMediaMessage.listRoomMediaMessages(intent));
+        List<RoomMediaMessage> cachedFiles = new ArrayList<>(RoomMediaMessage.listRoomMediaMessages(intent));
 
         if (null != cachedFiles) {
             for (RoomMediaMessage sharedDataItem : cachedFiles) {
@@ -142,7 +141,7 @@ public class VectorSharedFilesActivity extends RiotAppCompatActivity {
         if (0 != cachedFiles.size()) {
             Intent shareIntent = new Intent();
             shareIntent.setAction(Intent.ACTION_SEND_MULTIPLE);
-            shareIntent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, cachedFiles);
+            shareIntent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, (ArrayList) cachedFiles);
             shareIntent.setExtrasClassLoader(RoomMediaMessage.class.getClassLoader());
             shareIntent.setType("*/*");